% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Call TinySegmenter}
\usage{
segment(text, ...)
}
\arguments{
\item{text}{character vector}

\item{...}{all other args are passed to \code{jsonlite::fromJSON}}
}
\value{
list
}
\description{
Call TinySegmenter
}
