% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentrr-package.R
\docType{package}
\name{segmentrr-package}
\alias{segmentrr}
\alias{segmentrr-package}
\title{segmentrr: 'Rcpp' Port of TinySegmenter}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An R package for Japanese text tokenization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/segmentrr}
  \item Report bugs at \url{https://github.com/paithiov909/segmentrr}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Naoki Okamura
  \item Taku Kudo
}

}
\keyword{internal}
