% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentrr-package.R
\docType{package}
\name{segmentrr-package}
\alias{segmentrr}
\alias{segmentrr-package}
\title{segmentrr: segmentrr}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

One paragraph description of what the package does as one or more
         full sentences.
}
\keyword{internal}
